changed int to long long int
faced problem in leetcode contest and long long resolved the overflow problem so changed

+++ modular_arthimetic.cpp
@@ -5,8 +5,8 @@
 #define mod 1e9 + 7;
 using namespace std;
 
-int expo(int a, int b, int mod) {
-    int res = 1;
+int expo(long long int a, long long int b, long long int mod) {
+    long long int res = 1;
     while (b > 0) {
         if (b & 1)res = (res * a) % mod; 
         a = (a * a) % mod; 
@@ -15,21 +15,21 @@ int expo(int a, int b, int mod) {
     return res;
 }
 
-int mminvprime(int a, int b) {return expo(a, b - 2, b);}
+int mminvprime(long long int a, long long int b) {return expo(a, b - 2, b);}
 // addition operation
-int mod_add(int a, int b, int m) {
+int mod_add(long long int a, long long int b, long long int m) {
     a = a % m; b = b % m; return (((a + b) % m) ) % m; }
 //multiplication
-int mod_mul(int a, int b, int m) {
+int mod_mul(long long int a, long long int b, long long int m) {
     a = a % m; b = b % m; return (((a * b) % m) ) % m; }
 // subtration
-int mod_sub(int a, int b, int m){
+int mod_sub(long long int a, long long int b, long long int m){
     a = a % m; b = b % m; return (((a - b) % m) + m) % m; }
 // division
 //fermit little theorm
     // (a^p-a)%p==0 where a is any number and
     // p is any prime number
-int mod_div(int a, int b, int m) {
+int mod_div(long long int a, long long int b, long long int m) {
     a = a % m; b = b % m; return (mod_mul(a, mminvprime(b, m), m) ) % m;}  //only for prime m
 signed main(){
